Add a 'like' feature for UI
Users can now 'like' individual photos by clicking on the heart-icon next to the tags.  Liking a photo will call the like service, which would send a request to the API to like the photo on the website.  However the actual request has not yet been implemented as the current focus of this feature is on the UI development.